+++ backends/billing/src/main/java/com/arm/mbed/cloud/sdk/internal/billing/api/DefaultApi.java
@@ -1,28 +1,12 @@
 package com.arm.mbed.cloud.sdk.internal.billing.api;
 
-import com.arm.mbed.cloud.sdk.internal.billing.CollectionFormats.*;
-
-import retrofit2.Call;
-import retrofit2.http.*;
-
-import okhttp3.RequestBody;
-import okhttp3.ResponseBody;
-import okhttp3.MultipartBody;
-
-import com.arm.mbed.cloud.sdk.internal.billing.model.BadRequestErrorResponse;
-import com.arm.mbed.cloud.sdk.internal.billing.model.ForbiddenErrorResponse;
-import com.arm.mbed.cloud.sdk.internal.billing.model.InternalServerErrorResponse;
-import com.arm.mbed.cloud.sdk.internal.billing.model.ReportNotFoundErrorResponse;
 import com.arm.mbed.cloud.sdk.internal.billing.model.ReportResponse;
 import com.arm.mbed.cloud.sdk.internal.billing.model.ServicePackageQuota;
 import com.arm.mbed.cloud.sdk.internal.billing.model.ServicePackageQuotaHistoryResponse;
 import com.arm.mbed.cloud.sdk.internal.billing.model.ServicePackagesResponse;
-import com.arm.mbed.cloud.sdk.internal.billing.model.UnauthorizedErrorResponse;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import retrofit2.Call;
+import retrofit2.http.GET;
 
 public interface DefaultApi {
   /**
@@ -31,8 +15,30 @@ public interface DefaultApi {
    * @param month Queried year and month of billing report (required)
    * @return Call&lt;ReportResponse&gt;
    */
-  @GET("billing-report")
+  @GET("v3/billing-report")
   Call<ReportResponse> getBillingReport(
+    @retrofit2.http.Query("month") String month
+  );
+
+  /**
+   * Get raw active devices billing data for the month.
+   * Fetch raw active devices billing data for the currently authenticated commercial non-subtenant account. They are supplementary data for billing report. The raw active devices billing data for subtenant accounts are included in their aggregator&#39;s raw active devices billing data.
+   * @param month Queried year and month of billing report (required)
+   * @return Call&lt;Void&gt;
+   */
+  @GET("v3/billing-report-active-devices")
+  Call<Void> getBillingReportActiveDevices(
+    @retrofit2.http.Query("month") String month
+  );
+
+  /**
+   * Get raw firmware updates billing data for the month.
+   * Fetch generated firmware update devices billing report for the currently authenticated commercial non-subtenant account. The firmware update devices billing reports for subtenant accounts are included in their aggregator&#39;s firmware update devices billing report.
+   * @param month Queried year and month of billing report (required)
+   * @return Call&lt;Void&gt;
+   */
+  @GET("v3/billing-report-firmware-updates")
+  Call<Void> getBillingReportFirmwareUpdates(
     @retrofit2.http.Query("month") String month
   );
 
@@ -41,7 +47,7 @@ public interface DefaultApi {
    * Get the available firmware update quota for the currently authenticated commercial acount.
    * @return Call&lt;ServicePackageQuota&gt;
    */
-  @GET("service-packages-quota")
+  @GET("v3/service-packages-quota")
   Call<ServicePackageQuota> getServicePackageQuota();
     
 
@@ -52,7 +58,7 @@ public interface DefaultApi {
    * @param after To fetch after which quota history id. The results will contain entries after specified entry. (optional)
    * @return Call&lt;ServicePackageQuotaHistoryResponse&gt;
    */
-  @GET("service-packages-quota-history")
+  @GET("v3/service-packages-quota-history")
   Call<ServicePackageQuotaHistoryResponse> getServicePackageQuotaHistory(
     @retrofit2.http.Query("limit") Integer limit, @retrofit2.http.Query("after") String after
   );
@@ -62,7 +68,7 @@ public interface DefaultApi {
    * Get information of all service packages for currently authenticated commercial account. The response is returned with descending order by service package created timestamp, listing first pending service package, then active service package, and previous service packages at last.
    * @return Call&lt;ServicePackagesResponse&gt;
    */
-  @GET("service-packages")
+  @GET("v3/service-packages")
   Call<ServicePackagesResponse> getServicePackages();